requirements.txt
asgiref==3.3.0
astroid==2.4.2
colorama==0.4.4
Django==3.1.3
django-cors-headers==3.5.0
djangorestframework==3.12.1
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==8.0.1
pylint==2.6.0
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
wrapt==1.12.1
